+++ hvega/CHANGELOG.md
@@ -17,9 +17,11 @@ limited).
 The `HTitleFontWeight` constructor (a member of `HeaderProperty`)
 now takes a `FontWeight` argument rather than `Text`.
 
-The `LeTitle` constructor from `LegendConfig` was removed as it
-is not supported in Vega-Lite (`LeNoTitle` remains, as it is used
-to remove legend titles from a visualization).
+The `LeTitle` constructor from `LegendConfig` was removed as it is not
+supported in Vega-Lite (`LeNoTitle` remains, as it is used to remove
+legend titles from a visualization).
+
+`ScBinLinear` was removed from `Scale` as it is not used by Vega-Lite.
 
 ### New constructors